azure_xia/requirements.txt
python-dotenv
pandas
tqdm
tenacity
requests
biopython
